selected functionality [hit a wall]
I'm out of time today - have to pack for holiday.

I may return to this, mainly for practice and continue my learning

currently I'm snagged on moving the form value into the JS I've written in an attempt to make a dynamic href - my HTML and JS is a little rusty.